Create implementation
To convert the method to use a lambda function, you'll need to use a forEach loop with a lambda expression. Here's how you can do it in Java:
ambda functions in Java can be used with various functional interfaces, not just with the forEach method. Here are a few common places where lambda functions are often used:

Collections Framework:

forEach for iterating.
removeIf for removing elements.
sort for sorting.
map and reduce operations in streams.
Functional Interfaces:

Runnable for creating threads.
Comparator for comparing objects.
Callable for tasks that return a result.
Custom functional interfaces defined with a single abstract metho